adjust(client): product reactivity
